+++ src/client/stylesheets/main.scss
@@ -387,11 +387,24 @@ text.recharts-text.recharts-cartesian-axis-tick-value {
   padding: 0 15px;
 }
 
-.top-menu-drawer {
-  top: 80px !important;
-  position: fixed;
-  width: 100vw;
-  height: calc(100vh - 80px);
-  left: 0;
-  z-index: 999;
+.top-nav-submenu > .ant-menu.ant-menu-sub.ant-menu-vertical {
+  border: 0;
+  border-radius: 0 !important;
+  background-color: #262632 !important;
+  padding-top: 10px !important;
+  text-transform: capitalize !important;
+}
+
+#top-mobile-menu-nav .ant-drawer-content,
+#top-mobile-menu-nav .ant-menu-sub.ant-menu-inline {
+  background-color: #262632 !important;
+  box-shadow: none;
+}
+
+.language-switcher ul {
+  background-color: #262632 !important;
+  padding: 20px !important;
+  padding-top: 30px !important;
+  min-width: 150px !important;
+  margin-left: -20px !important;
 }
